search/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bandit==1.7.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
cryptography==37.0.4
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.1
django-crispy-forms==1.14.0
e==1.4.5
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.2
Flask==2.2.2
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.24
greenlet==2.0.0a2
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.12.0
importlib-resources==5.4.0
ipdb==0.13.9
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
memory-profiler==0.60.0
missingno==0.5.1
mistune==0.8.4
more-itertools==8.12.0
multimethod==1.8
mysql-client==0.0.1
mysql-connector-python==8.0.17
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8.3
notebook==6.4.7
numpy==1.22.4
openpyxl==3.0.9
packaging==21.3
panda==0.3.1
pandas==1.4.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
py4j==0.10.9.2
pyappleemail==1.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.11.1
PyMySQL==1.0.2
pyobjc==8.1
pyobjc-core==8.1
pyobjc-framework-Accessibility==8.1
pyobjc-framework-Accounts==8.1
pyobjc-framework-AddressBook==8.1
pyobjc-framework-AdServices==8.1
pyobjc-framework-AdSupport==8.1
pyobjc-framework-AppleScriptKit==8.1
pyobjc-framework-AppleScriptObjC==8.1
pyobjc-framework-ApplicationServices==8.1
pyobjc-framework-AppTrackingTransparency==8.1
pyobjc-framework-AudioVideoBridging==8.1
pyobjc-framework-AuthenticationServices==8.1
pyobjc-framework-AutomaticAssessmentConfiguration==8.1
pyobjc-framework-Automator==8.1
pyobjc-framework-AVFoundation==8.1
pyobjc-framework-AVKit==8.1
pyobjc-framework-BusinessChat==8.1
pyobjc-framework-CalendarStore==8.1
pyobjc-framework-CallKit==8.1
pyobjc-framework-CFNetwork==8.1
pyobjc-framework-ClassKit==8.1
pyobjc-framework-CloudKit==8.1
pyobjc-framework-Cocoa==8.1
pyobjc-framework-Collaboration==8.1
pyobjc-framework-ColorSync==8.1
pyobjc-framework-Contacts==8.1
pyobjc-framework-ContactsUI==8.1
pyobjc-framework-CoreAudio==8.1
pyobjc-framework-CoreAudioKit==8.1
pyobjc-framework-CoreBluetooth==8.1
pyobjc-framework-CoreData==8.1
pyobjc-framework-CoreHaptics==8.1
pyobjc-framework-CoreLocation==8.1
pyobjc-framework-CoreMedia==8.1
pyobjc-framework-CoreMediaIO==8.1
pyobjc-framework-CoreMIDI==8.1
pyobjc-framework-CoreML==8.1
pyobjc-framework-CoreMotion==8.1
pyobjc-framework-CoreServices==8.1
pyobjc-framework-CoreSpotlight==8.1
pyobjc-framework-CoreText==8.1
pyobjc-framework-CoreWLAN==8.1
pyobjc-framework-CryptoTokenKit==8.1
pyobjc-framework-DeviceCheck==8.1
pyobjc-framework-DictionaryServices==8.1
pyobjc-framework-DiscRecording==8.1
pyobjc-framework-DiscRecordingUI==8.1
pyobjc-framework-DiskArbitration==8.1
pyobjc-framework-DVDPlayback==8.1
pyobjc-framework-EventKit==8.1
pyobjc-framework-ExceptionHandling==8.1
pyobjc-framework-ExecutionPolicy==8.1
pyobjc-framework-ExternalAccessory==8.1
pyobjc-framework-FileProvider==8.1
pyobjc-framework-FileProviderUI==8.1
pyobjc-framework-FinderSync==8.1
pyobjc-framework-FSEvents==8.1
pyobjc-framework-GameCenter==8.1
pyobjc-framework-GameController==8.1
pyobjc-framework-GameKit==8.1
pyobjc-framework-GameplayKit==8.1
pyobjc-framework-ImageCaptureCore==8.1
pyobjc-framework-IMServicePlugIn==8.1
pyobjc-framework-InputMethodKit==8.1
pyobjc-framework-InstallerPlugins==8.1
pyobjc-framework-InstantMessage==8.1
pyobjc-framework-Intents==8.1
pyobjc-framework-IOSurface==8.1
pyobjc-framework-iTunesLibrary==8.1
pyobjc-framework-KernelManagement==8.1
pyobjc-framework-LatentSemanticMapping==8.1
pyobjc-framework-LaunchServices==8.1
pyobjc-framework-libdispatch==8.1
pyobjc-framework-LinkPresentation==8.1
pyobjc-framework-LocalAuthentication==8.1
pyobjc-framework-MapKit==8.1
pyobjc-framework-MediaAccessibility==8.1
pyobjc-framework-MediaLibrary==8.1
pyobjc-framework-MediaPlayer==8.1
pyobjc-framework-MediaToolbox==8.1
pyobjc-framework-Metal==8.1
pyobjc-framework-MetalKit==8.1
pyobjc-framework-MetalPerformanceShaders==8.1
pyobjc-framework-MetalPerformanceShadersGraph==8.1
pyobjc-framework-MLCompute==8.1
pyobjc-framework-ModelIO==8.1
pyobjc-framework-MultipeerConnectivity==8.1
pyobjc-framework-NaturalLanguage==8.1
pyobjc-framework-NetFS==8.1
pyobjc-framework-Network==8.1
pyobjc-framework-NetworkExtension==8.1
pyobjc-framework-NotificationCenter==8.1
pyobjc-framework-OpenDirectory==8.1
pyobjc-framework-OSAKit==8.1
pyobjc-framework-OSLog==8.1
pyobjc-framework-PassKit==8.1
pyobjc-framework-PencilKit==8.1
pyobjc-framework-Photos==8.1
pyobjc-framework-PhotosUI==8.1
pyobjc-framework-PreferencePanes==8.1
pyobjc-framework-PushKit==8.1
pyobjc-framework-Quartz==8.1
pyobjc-framework-QuickLookThumbnailing==8.1
pyobjc-framework-ReplayKit==8.1
pyobjc-framework-SafariServices==8.1
pyobjc-framework-SceneKit==8.1
pyobjc-framework-ScreenSaver==8.1
pyobjc-framework-ScreenTime==8.1
pyobjc-framework-ScriptingBridge==8.1
pyobjc-framework-SearchKit==8.1
pyobjc-framework-Security==8.1
pyobjc-framework-SecurityFoundation==8.1
pyobjc-framework-SecurityInterface==8.1
pyobjc-framework-ServiceManagement==8.1
pyobjc-framework-Social==8.1
pyobjc-framework-SoundAnalysis==8.1
pyobjc-framework-Speech==8.1
pyobjc-framework-SpriteKit==8.1
pyobjc-framework-StoreKit==8.1
pyobjc-framework-SyncServices==8.1
pyobjc-framework-SystemConfiguration==8.1
pyobjc-framework-SystemExtensions==8.1
pyobjc-framework-UniformTypeIdentifiers==8.1
pyobjc-framework-UserNotifications==8.1
pyobjc-framework-UserNotificationsUI==8.1
pyobjc-framework-VideoSubscriberAccount==8.1
pyobjc-framework-VideoToolbox==8.1
pyobjc-framework-Virtualization==8.1
pyobjc-framework-Vision==8.1
pyobjc-framework-WebKit==8.1
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
stevedore==3.5.0
tabulate==0.8.9
tangled-up-in-unicode==0.2.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.13.1
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.2
yapf==0.32.0
zipp==3.7.0
